requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
batchgenerators==0.25.1
beautifulsoup4==4.12.3
Brotli==1.0.9
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
dm-tree==0.1.8
docker-pycreds==0.4.0
dominate==2.9.1
einops==0.7.0
eval_type_backport==0.2.0
ffmpeg==1.4
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2024.10.0
future==0.18.3
gast==0.5.4
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.41
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
huggingface-hub==0.26.1
idna==3.4
imageio==2.33.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keras-nightly==3.0.4.dev2024021403
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
libtiff==0.4.2
linecache2==1.0.0
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.2
mdurl==0.1.2
MedPy==0.5.2
meson==1.8.3
mindspore==2.3.1
mkl_fft==1.3.10
mkl_random==1.2.8
mkl-service==2.4.1
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-contrib-python==4.10.0.82
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pathtools==0.1.2
Pillow==9.3.0
pip==23.3.1
platformdirs==4.2.0
pooch==1.8.2
promise==2.3
protobuf==4.25.2
psutil==5.9.8
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-msssim==1.0.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
rawpy==0.25.0
requests==2.31.0
rich==13.7.0
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.10.1
sentry-sdk==1.40.1
setproctitle==1.3.3
setuptools==68.0.0
shortuuid==1.0.13
SimpleITK==2.4.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.13.3
tb-nightly==2.16.0a20240212
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras-nightly==2.17.0.dev2024021422
tf-nightly==2.17.0.dev20240214
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2023.9.26
timm==1.0.11
tomli==2.0.1
torch==2.7.1+cu118
torchaudio==0.12.1
torchsummary==1.5.1
torchvision==0.22.1+cu118
tornado==6.4
tqdm==4.64.1
traceback2==1.4.0
triton==3.3.1
typing_extensions==4.12.2
tzdata==2023.4
unittest2==1.1.0
urllib3==1.26.18
utils==1.0.2
visdom==0.2.4
wandb==0.13.5
warmup-scheduler==0.3
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0
